モニター設定のフォーマット修正
