--- F0V2/Util/MonitorConfiguration.h
@@ -1,7 +1,0 @@
-#pragma once
-
-inline namespace Util_inline
-{
-    [[nodiscard]]
-    std::string SerializeCurrentMonitorConfiguration();
-}
