+++ F0V2/Util/MonitorInfo.h
@@ -1,0 +1,13 @@
+#pragma once
+
+inline namespace Util_inline
+{
+    struct MonitorInfo
+    {
+        std::string deviceName;
+        RECT rect{};
+        bool primary{};
+    };
+
+    std::vector<MonitorInfo> EnumerateMonitors();
+}
